+++ cmd/wacli/root.go
@@ -79,6 +79,7 @@ func execute(args []string) error {
 	rootCmd.AddCommand(newAuthCmd(&flags))
 	rootCmd.AddCommand(newSyncCmd(&flags))
 	rootCmd.AddCommand(newMessagesCmd(&flags))
+	rootCmd.AddCommand(newChangesCmd(&flags))
 	rootCmd.AddCommand(newCallsCmd(&flags))
 	rootCmd.AddCommand(newSendCmd(&flags))
 	rootCmd.AddCommand(newPollCmd(&flags))